Extract emitAddMem(Address,int,codeGen&)
