+++ dyninstAPI/CMakeLists.txt
@@ -270,6 +270,8 @@ if(DYNINST_CODEGEN_ARCH_I386 OR DYNINST_CODEGEN_ARCH_X86_64)
     _sources
     src/codegen/emitters/x86/Emitterx86.C
     src/codegen/emitters/x86/Emitterx86.h
+    src/codegen/emitters/x86/generators.C
+    src/codegen/emitters/x86/generators.h
     src/RegisterConversion-x86.C
     src/registerSpace/registerSpace-x86.C
     src/Relocation/Widgets/CFWidget-x86.C
